Support deleting ThemeManager after actiongroup and menu, to fix crash in kmail
Summary:
We used to leak the ThemeManager. Now that we delete it, it crashed
because we delete the action group and the menu first. Add support for
setActionGroup(nullptr) and setThemeMenu(nullptr) so we can clean up while
everything still exists.

Test Plan:
Unittest. The actual fix for kmail requires calling these methods
in grantleeheaderstyleplugin.

Reviewers: mlaurent

Reviewed By: mlaurent

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D12574