Grantleetheme should always prefer files that are located in DataLocation.
Summary:
Tests need somehow provide the current files and not rely on, that these
files are already installed (see D15727). Having no special
ordering in place, it is random, what file is picked, when there are
themes with same names on different locations. So we may need sort the
themesDirectories.

Reviewers: #kde_pim, mlaurent, dfaure

Reviewed By: dfaure

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D17954